+++ src/main/java/com/project/devidea/modules/account/controllers/CommonSignUpController.java
@@ -2,10 +2,12 @@ package com.project.devidea.modules.account.controllers;
 
 import com.project.devidea.infra.error.GlobalResponse;
 import com.project.devidea.modules.account.dto.SignUp;
-import com.project.devidea.modules.account.services.signUp.SignUpAndAuthenticationEmailService;
+import com.project.devidea.modules.account.services.signUp.AuthenticationEmailService;
+import com.project.devidea.modules.account.services.signUp.SignUpService;
 import com.project.devidea.modules.account.validator.SignUpRequestValidator;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.WebDataBinder;
@@ -15,14 +17,13 @@ import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.io.IOException;
 
-/**
- * 일반 회원가입, 메일인증, 회원탈퇴 로직
- */
 @RestController
 @RequiredArgsConstructor
 public class CommonSignUpController {
 
-    private final SignUpAndAuthenticationEmailService signUpService;
+    @Qualifier("commonSignUpService")
+    private final SignUpService signUpService;
+    private final AuthenticationEmailService authenticationEmailService;
     private final SignUpRequestValidator signUpRequestValidator;
 
     @InitBinder("commonRequest")
@@ -41,7 +42,7 @@ public class CommonSignUpController {
     @ApiOperation("일반 회원가입 메일 인증")
     public void authenticateEmailToken(String email, String token, HttpServletResponse response) throws IOException {
 
-        String url = signUpService.authenticateEmailToken(email, token);
+        String url = authenticationEmailService.authenticateEmailToken(email, token);
         response.sendRedirect(url + "/sign-up/detail?token=" + token);
     }
 }
