+++ src/main/java/com/project/devidea/modules/account/services/signUp/OAuthSignUpServiceImpl.java
@@ -4,6 +4,7 @@ import com.project.devidea.modules.account.domains.Account;
 import com.project.devidea.modules.account.dto.SignUp;
 import com.project.devidea.modules.account.repository.AccountRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
+@Qualifier("oAuthSignUpService")
 public class OAuthSignUpServiceImpl implements SignUpService {
 
     private final AccountRepository accountRepository;
