+++ src/test/java/com/project/devidea/modules/account/controllers/CommonSignUpControllerValidationTest.java
@@ -1,0 +1,59 @@
+package com.project.devidea.modules.account.controllers;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.project.devidea.modules.account.AccountDummy;
+import com.project.devidea.modules.account.dto.SignUp;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.hamcrest.Matchers.is;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@WebAppConfiguration
+@AutoConfigureMockMvc
+public class CommonSignUpControllerValidationTest {
+
+    @Autowired MockMvc mockMvc;
+    @Autowired ObjectMapper objectMapper;
+
+    @Test
+    void 일반회원가입_기본Valid() throws Exception {
+
+//        given
+        SignUp.CommonRequest failValidSignUpRequest = AccountDummy.getFailSignUpRequestWithValid();
+
+//        when, then
+//        이메일 공백, password Blank, passwordConfirm Blank, name 길이, nickname Blank
+        mockMvc.perform(post("/sign-up")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(failValidSignUpRequest)))
+                .andDo(print())
+                .andExpect(status().is4xxClientError())
+                .andExpect(jsonPath("$.errors.length()", is(5)));
+    }
+
+    @Test
+    void 일반회원가입_WithValidator() throws Exception {
+
+//        given
+        SignUp.CommonRequest commonRequest = AccountDummy.getFailSignUpRequestWithValidator();
+
+//        when, then
+//        이메일 중복과 패스워드, 패스워드 확인값 불일치, 닉네임 중복
+        mockMvc.perform(post("/sign-up")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(commonRequest)))
+                .andDo(print())
+                .andExpect(status().is4xxClientError())
+                .andExpect(jsonPath("$.errors.length()", is(3)));
+    }
+}
